deploy/requirements.txt
autopep8==1.5.3
certifi==2020.4.5.2
chardet==3.0.4
gitdb==4.0.5
GitPython==3.1.3
idna==2.9
pycodestyle==2.6.0
requests==2.23.0
smmap==3.0.4
toml==0.10.1
urllib3==1.25.9
